Error when `gh repo rename` is used with a new repo name that contains an owner
`gh repo rename` can only rename a repo, but not change its owner
(i.e. transfer it).

As flagged in #10034, at the moment, the CLI behaves like it can do
this, and produces weird results.

If I ask to rename my repo to `polyseam/frappe-containers`:

```bash
gh repo rename polyseam/frappe-containers
```

..the preview suggests that it will do the right thing:

```
? Rename polyseam/containers to polyseam/frappe-containers? Yes
```

...but the repo gets renamed wrongly:

```
✓ Renamed repository polyseam/polyseam-frappe-containers
```

This adds an upfront validation that looks for a slash and tells
you that the command can't transfer a repo.

Fixes #10034.